+++ CHANGELOG.md
@@ -8,6 +8,12 @@ The format is based on [Keep a Changelog][keepachangelog] and this project adher
 
 ### Added
 
+- Integration with [Laravel Telescope](https://github.com/laravel/telescope/) is supported now (just enable `SetupTelescopeListener` for `BeforeLoopStartedEvent`)
+
+## v5.1.0
+
+### Added
+
 - Listener `FlushLogContextListener` for the logger context flushing
 
 ## v5.0.2